+++ arr_equal_test.go
